# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sturtz Maschinenbau Support Chatbot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/documents:
    post:
      operationId: registerDocument
      tags: [admin]
      summary: Register a document after upload (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDocumentBody"
      responses:
        "201":
          description: Document registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /admin/documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [admin]
      summary: Delete a document (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/documents/{id}/reingest:
    post:
      operationId: reingestDocument
      tags: [admin]
      summary: Trigger re-ingestion of a document (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Re-ingestion triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}/view:
    get:
      operationId: getDocumentViewUrl
      tags: [documents]
      summary: Get a redirect URL to view a document at a specific page
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          required: false
          schema:
            type: integer
      responses:
        "302":
          description: Redirect to file
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        filename:
          type: string
        filePath:
          type: string
        fileType:
          type: string
        size:
          type: integer
        status:
          type: string
          enum: [pending, ingesting, ready, failed]
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
      required: [id, title, filename, filePath, fileType, size, status, uploadedBy, createdAt]
    RegisterDocumentBody:
      type: object
      properties:
        title:
          type: string
        filename:
          type: string
        objectPath:
          type: string
        fileType:
          type: string
        size:
          type: integer
      required: [title, filename, objectPath, fileType, size]
...