# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sturtz Maschinenbau Support Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with email and password
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.string().uuid(),
    email: zod.string(),
    name: zod.string(),
    role: zod.enum(["admin", "user"]),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get the current authenticated user
 */
export const GetCurrentUserResponse = zod.object({
  id: zod.string().uuid(),
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["admin", "user"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all users (admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string().uuid(),
  email: zod.string(),
  name: zod.string(),
  role: zod.enum(["admin", "user"]),
  createdAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a new user (admin only)
 */
export const CreateUserBody = zod.object({
  email: zod.string(),
  name: zod.string(),
  password: zod.string(),
  role: zod.enum(["admin", "user"]),
});

/**
 * @summary Delete a user (admin only)
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.string().uuid(),
});

/**
 * @summary Request a presigned URL for direct GCS upload
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
  metadata: zod.object({
    name: zod.string(),
    size: zod.number(),
    contentType: zod.string(),
  }),
});

/**
 * @summary List all documents in the knowledge base
 */
export const ListDocumentsResponseItem = zod.object({
  id: zod.string().uuid(),
  title: zod.string(),
  filename: zod.string(),
  filePath: zod.string(),
  fileType: zod.string(),
  size: zod.number(),
  status: zod.enum(["pending", "ingesting", "ready", "failed"]),
  uploadedBy: zod.string().uuid(),
  createdAt: zod.coerce.date(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Register a document after upload (admin only)
 */
export const RegisterDocumentBody = zod.object({
  title: zod.string(),
  filename: zod.string(),
  objectPath: zod.string(),
  fileType: zod.string(),
  size: zod.number(),
});

/**
 * @summary Delete a document (admin only)
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.string().uuid(),
});

/**
 * @summary Trigger re-ingestion of a document (admin only)
 */
export const ReingestDocumentParams = zod.object({
  id: zod.coerce.string().uuid(),
});

/**
 * @summary Get a redirect URL to view a document at a specific page
 */
export const GetDocumentViewUrlParams = zod.object({
  id: zod.coerce.string().uuid(),
});

export const GetDocumentViewUrlQueryParams = zod.object({
  page: zod.coerce.number().optional(),
});

/**
 * @summary List the current user's conversations
 */
export const ListConversationsResponseItem = zod.object({
  id: zod.string().uuid(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListConversationsResponse = zod.array(
  ListConversationsResponseItem,
);

/**
 * @summary Start a new conversation
 */
export const CreateConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get a conversation with all messages
 */
export const GetConversationParams = zod.object({
  id: zod.coerce.string().uuid(),
});

export const GetConversationResponse = zod.object({
  id: zod.string().uuid(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.string().uuid(),
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
      citations: zod.array(
        zod.object({
          documentId: zod.string().uuid(),
          documentTitle: zod.string(),
          pageNumber: zod.number(),
          snippet: zod.string(),
        }),
      ),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteConversationParams = zod.object({
  id: zod.coerce.string().uuid(),
});

/**
 * @summary Send a user message and get an assistant reply (proxies to RAG service)
 */
export const SendMessageParams = zod.object({
  id: zod.coerce.string().uuid(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
});

export const SendMessageResponse = zod.object({
  userMessage: zod.object({
    id: zod.string().uuid(),
    role: zod.enum(["user", "assistant"]),
    content: zod.string(),
    citations: zod.array(
      zod.object({
        documentId: zod.string().uuid(),
        documentTitle: zod.string(),
        pageNumber: zod.number(),
        snippet: zod.string(),
      }),
    ),
    createdAt: zod.coerce.date(),
  }),
  assistantMessage: zod.object({
    id: zod.string().uuid(),
    role: zod.enum(["user", "assistant"]),
    content: zod.string(),
    citations: zod.array(
      zod.object({
        documentId: zod.string().uuid(),
        documentTitle: zod.string(),
        pageNumber: zod.number(),
        snippet: zod.string(),
      }),
    ),
    createdAt: zod.coerce.date(),
  }),
});