# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sturtz Maschinenbau Support Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: string;
  title: string;
  filename: string;
  filePath: string;
  fileType: string;
  size: number;
  status: DocumentStatus;
  uploadedBy: string;
  createdAt: Date;
}